style: move order levels and sets into an orders file antara::orders::
